#222 - fix full issue
The problem was that guidelines checked for the original target date, while the UI received a target date that was modified to ensure that, in this case, the target date of influenza fell between october and march. Solution changes the target date that the guidelines see, ensuring that all parts see the same target date.